+++ src/htmlcomb.js
@@ -1,13 +1,36 @@
+/*!
+ * HTMLComb v@VERSION
+ * https://github.com/fengyuanchen/htmlcomb
+ *
+ * Copyright 2014 Fengyuan Chen
+ * Released under the MIT license
+ *
+ * Date: @DATE
+ */
+
 (function () {
 
-  "use strict";
+  'use strict';
 
   var HTMLComb = function (options) {
         this.defaults = util.extend({}, HTMLComb.DEFAULTS, options);
-        this.source = "";
-        this.result = "";
+        this.source = '';
+        this.result = '';
       },
 
+      // Patterns
+      REGEXP_ELEMENT_TAGS = /<(\w+)\s([^<>]+)>/g,
+      REGEXP_ELEMENT_ATTRIBUTES = /\w+(?:\-\w*)*(\=(("[^"]*")|('[^']*')|\w*))?/g,
+      REGEXP_SINGLE_QUOTATION_MARKS = /'([^']*)'/g,
+      REGEXP_DOUBLE_QUOTATION_MARKS = /"/g,
+      REGEXP_EMPTY_VALUES = /^"\s*"$/,
+      REGEXP_NEWLINES = /[\n\r]+/g,
+      REGEXP_MULTIPLE_SPACES = /(\s)\s+/g,
+
+      // Others
+      array = [],
+      slice = array.slice,
+      toString = {}.toString,
       util;
 
   // Defaults
@@ -16,21 +39,23 @@
   HTMLComb.DEFAULTS = {
     requireDoubleQuotationMarks: true,
     replaceSingleQuotationMarks: true,
-    removeEmptyValue: true,
+    removeMultipleSpaces: true,
+    removeEmptyValues: true,
+    removeNewlines: true,
     order: [
-      "class",
-      "id",
-      "name",
-      "data",
-      "src",
-      "for",
-      "type",
-      "href",
-      "value",
-      "title",
-      "alt",
-      "aria",
-      "role"
+      'class',
+      'id',
+      'name',
+      'data',
+      'src',
+      'for',
+      'type',
+      'href',
+      'value',
+      'title',
+      'alt',
+      'aria',
+      'role'
     ]
   };
 
@@ -41,33 +66,27 @@
     constructor: HTMLComb,
 
     setup: function (options) {
-      if (typeof options === "object") {
+      if (typeof options === 'object') {
         util.extend(this.defaults, options);
       }
     },
 
-    format: function (source, options, callback) {
+    format: function (source, callback) {
       var result;
 
-      if (typeof source !== "string") {
-        throw new Error("The first parameter for `format` method must be a string.");
+      if (typeof source !== 'string') {
+        throw new Error('The first parameter for "format" method must be a string.');
       }
 
       this.source = source;
 
-      if (typeof options === "object") {
-        util.extend(this.defaults, options);
-      } else if (typeof options === "function" && typeof callback === "undefined") {
-        callback = options;
-      }
-
-      result = source.replace(/<(\w+)\s([^<>]+)>/g, util.proxy(function (tag, tagName, attrs) {
-        return ("<" + tagName + " " + this.sort(attrs) + ">");
+      result = source.replace(REGEXP_ELEMENT_TAGS, util.proxy(function (tag, tagName, attrs) {
+        return ('<' + tagName + ' ' + this.sort(attrs) + '>');
       }, this));
 
       this.result = result;
 
-      if (typeof callback === "function") {
+      if (typeof callback === 'function') {
         callback.call(this, result);
       }
 
@@ -93,7 +112,7 @@
           if (attr.substr(0, attrName.length) === attrName) {
             matched = true;
 
-            if (typeof matchedAttrs[i] === "undefined") {
+            if (typeof matchedAttrs[i] === 'undefined') {
               matchedAttrs[i] = [];
             }
 
@@ -110,51 +129,59 @@
 
       // Filters undefined values and sorts matched attrs
       util.each(matchedAttrs, function (attr) {
-        if (Array.isArray(attr)) {
+        if (util.isArray(attr)) {
 
           if (attr.length > 1) {
             attr.sort();
           }
 
-          sortedAttrs.push(attr.join(" "));
+          sortedAttrs.push(attr.join(' '));
         }
       });
 
       attrs = sortedAttrs.concat(others.sort());
 
-      return attrs.join(" ");
+      return attrs.join(' ');
     },
 
     split: function (attrs) {
       var defaults = this.defaults,
           matched = [];
 
-      if (typeof attrs === "string") {
+      if (typeof attrs === 'string') {
 
         // Matchs four types of attribute: attr="prop" | attr='prop' | attr=prop | attr
-        matched = attrs.match(/\w+(?:\-\w*)*(\=(("[^"]*")|('[^']*')|\w*))?/g);
+        matched = attrs.match(REGEXP_ELEMENT_ATTRIBUTES);
 
         if (matched) {
           util.each(matched, function (attr, i) {
             var firstLetter;
 
-            attr = attr.split("=");
+            attr = attr.split('=');
 
-            if (typeof attr[1] !== "undefined") {
+            if (typeof attr[1] !== 'undefined') {
               firstLetter = attr[1].charAt(0);
 
-              if (firstLetter === "'" && defaults.replaceSingleQuotationMarks) { // Replases ' to "
-                attr[1] = '"' + attr[1].replace(/'([^']*)'/g, "$1").replace(/"/g, "&quot;") + '"';
+              if (firstLetter === '\'' && defaults.replaceSingleQuotationMarks) { // Replases ' to "
+                attr[1] = '"' + attr[1].replace(REGEXP_SINGLE_QUOTATION_MARKS, '$1').replace(REGEXP_DOUBLE_QUOTATION_MARKS, '&quot;') + '"';
               } else if (firstLetter !== '"' && defaults.requireDoubleQuotationMarks) { // Adds "
                 attr[1] = '"' + attr[1] + '"';
               }
 
-              if (attr[1] === '""' && defaults.removeEmptyValue) { // Removes empty value
+              if (defaults.removeNewlines) { // Removes newlines first
+                attr[1] = attr[1].replace(REGEXP_NEWLINES, '');
+              }
+
+              if (defaults.removeMultipleSpaces) {
+                attr[1] = attr[1].replace(REGEXP_MULTIPLE_SPACES, '$1');
+              }
+
+              if (REGEXP_EMPTY_VALUES.test(attr[1]) && defaults.removeEmptyValues) {
                 attr.pop();
               }
             }
 
-            matched[i] = attr.join("=");
+            matched[i] = attr.join('=');
           });
         } else {
           matched = [];
@@ -170,22 +197,36 @@
   // ---------------------------------------------------------------------------
 
   util = {
-    toArray: function (obj, srart, end) {
-      return Array.prototype.slice.call(obj, srart, end);
+    isArray: Array.isArray || function (arr) {
+      return toString.call(arr) === '[object Array]';
+    },
+
+    toArray: function (obj, start, end) {
+      var args = [];
+
+      if (typeof start === 'number') {
+        args.push(start);
+
+        if (typeof end === 'number') {
+          args.push(end);
+        }
+      }
+
+      return slice.apply(obj, args);
     },
 
     each: function (obj, callback) {
       var length,
           i;
 
-      if (typeof callback === "function") {
-        if (Array.isArray(obj)) {
+      if (typeof callback === 'function') {
+        if (util.isArray(obj)) {
           for (i = 0, length = obj.length; i < length; i++) {
             if (callback.call(obj, obj[i], i) === false) {
               break;
             }
           }
-        } else if (typeof obj === "object") {
+        } else if (typeof obj === 'object') {
           for (i in obj) {
             if (obj.hasOwnProperty(i)) {
               if (callback.call(obj, obj[i], i) === false) {
@@ -236,17 +277,17 @@
   // Define and export
   // ---------------------------------------------------------------------------
 
-  if (typeof window !== "undefined") {
+  if (typeof window !== 'undefined') {
     window.HTMLComb = HTMLComb;
   }
 
-  if (typeof define === "function" && define.amd) {
-    define("htmlcomb", [], function () {
+  if (typeof define === 'function' && define.amd) {
+    define('htmlcomb', [], function () {
       return HTMLComb;
     });
   }
 
-  if (typeof module === "object") {
+  if (typeof module === 'object') {
     module.exports = HTMLComb;
   }